+++ tests/testthat/test-Logbound.R
@@ -27,6 +27,9 @@ test_that("Coxph strata_basic_single_CR_null log_bound", {
   expect_error(LikelihoodBound(coxres, df, curve_control, control = control, bad = FALSE))
   expect_error(LikelihoodBound(coxres, df, curve_control, control = control, norm = "bad"))
   expect_error(LikelihoodBound(coxres, df, curve_control = "bad"))
+  expect_error(LikelihoodBound(coxres, df, para_number = 1.1))
+  expect_error(LikelihoodBound(coxres, df, para_number = -1))
+  expect_error(LikelihoodBound(coxres, df, para_number = 100))
   #
   for (m in c(TRUE, FALSE)) {
     a_n <- c(-0.1, -0.1)
@@ -93,6 +96,9 @@ test_that("Poisson strata_single log_bound", {
   expect_error(LikelihoodBound(poisres, df, curve_control, control = control, bad = FALSE))
   expect_error(LikelihoodBound(poisres, df, curve_control, control = control, norm = "bad"))
   expect_error(LikelihoodBound(poisres, df, curve_control = "bad"))
+  expect_error(LikelihoodBound(poisres, df, para_number = 1.1))
+  expect_error(LikelihoodBound(poisres, df, para_number = -1))
+  expect_error(LikelihoodBound(poisres, df, para_number = 100))
   #
   for (m in c(TRUE, FALSE)) {
     a_n <- c(-0.1, -0.1)
@@ -100,6 +106,13 @@ test_that("Poisson strata_single log_bound", {
     curve_control <- list("manual" = m)
     expect_no_error(LikelihoodBound(poisres, df, curve_control, control = control, bisect = FALSE))
     expect_no_error(LikelihoodBound(poisres_s, df, curve_control, control = control))
+    zz <- file(paste(tempfile(), ".txt", sep = ""), open = "wt")
+    sink(zz)
+    sink(zz, type = "message")
+    print(e)
+    sink(type = "message")
+    sink(NULL)
+    close(zz)
   }
   for (m in c(TRUE, FALSE)) {
     model_control <- list("strata" = FALSE, "single" = TRUE, "log_bound" = TRUE, "manual" = m)